+++ src/engine/index.ts
@@ -24,8 +24,8 @@ export * from './Events';
 export * from './Label';
 export { FontStyle, FontUnit, TextAlign, BaseAlign } from './Graphics/FontCommon';
 export * from './EmitterType';
-export { Particle, ParticleTransform, ParticleArgs, ParticleEmitterArgs } from './Particles';
-export * from './ParticleEmitter';
+export { Particle, ParticleTransform, ParticleConfig as ParticleArgs, ParticleEmitterArgs } from './Particles/Particles';
+export * from './Particles/ParticleEmitter';
 export * from './Collision/Physics';
 export * from './Scene';